When looking at the performance test step-3, I realized that our implementation in the precondition_SSOR did not really exploit all possible instruction-level parallelism for modern out-of-order processors: Since we are bound to the order of addition, we should start the summation with those values in the `dst` vector that have been computed a long time ago, i.e., the ones with the highest index that come at the end of the row. Once we get close to the diagonal, an out-of-order processor might still be working on one or a few rows before the current one. If I did the tests correctly, this reduces the time for the `timing_step_3 solve` stage from 2.38 seconds to 2.25 seconds on my laptop (AMD Zen2, 4.2 GHz) or from 2.80 to 2.67 seconds on a server (Intel Ice Lake, 3.5 GHz). Not a big change, but still consistently better.