@kronbichler 2022-02-16
I looked at the assembly code of the most expensive function in the code and saw that it ran forward even though my feeling suggested we should not do so in the backward sweep. I had been doing similar micro-optimizations in some other contexts based on matrix-free computations and in the analysis of the invariant-domain Euler computations (going beyond step-69), so I was aware of some common problems. Then it was merely trying things out. Of course, the elephant in the room here is the fact that one cannot parallelize the SSOR sweep (without additional work), so the effect got only visible by a single core having enough memory bandwidth to run into other limits in the first place.